==> notebooks/02_validate_and_save.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

import pandas as pd
from pathlib import Path
import pyarrow.parquet as pq

month = 1
year = 2023
path = Path('..') / 'data' / 'raw' / f"citi_rides_{year}_{month:02}.parquet"

table = pq.read_table(path)
rides = table.to_pandas()
rides.head()

rides_cp = rides.copy()
rides_cp.info()

time_cols = ['started_at', 'ended_at']
for col in time_cols:
    rides_cp[col] = pd.to_datetime(rides_cp[col], errors='coerce')

rides_cp['duration'] = rides_cp['ended_at'] - rides_cp['started_at']
rides_cp.head()

rides_cp['duration'].describe().T

rides_cp["duration"].quantile(0)
rides_cp["duration"].quantile(0.01)
rides_cp["duration"].quantile(0.99)
rides_cp["duration"].quantile(0.999)

duration_filter = (rides_cp["duration"] > pd.Timedelta(0)) & (rides_cp["duration"] <= pd.Timedelta(hours=1))
sum(~duration_filter)

sorted_df = rides_cp.sort_values(by="started_at", ascending=True)  

# Get the top 10 (smallest) and bottom 10 (largest) values  
top_10 = sorted_df.head(10)  
bottom_10 = sorted_df.tail(10)  

top_10

bottom_10

rides_cp['start_station_name'].nunique()

# Data in the file that should not be there, remove all values that should not be there
filter_date_range = (rides_cp['started_at'] >= '2023-01-01') & (rides_cp['started_at'] < '2023-02-01')
sum(~filter_date_range)

final_filter = duration_filter & filter_date_range
numbers_dropped = final_filter.shape[0] - sum(final_filter)
numbers_dropped
numbers_dropped / final_filter.shape[0] * 100

rides = rides[final_filter]
rides = rides[['started_at', 'start_station_id']]
rides.sort_values(by="started_at", ascending=True).head()

path = Path('..') / 'data' / 'processed' / f'citi_rides_{year}_{month:02}.parquet'
rides.to_parquet(path, engine="pyarrow", index=False)
