==> notebooks/07_baseline_models.ipynb <==
%load_ext autoreload
%autoreload 2

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

import pandas as pd

from src.config import PROCESSED_DATA_DIR

ts_data = pd.read_parquet(PROCESSED_DATA_DIR / "ts_data.parquet")
ts_data

ts_data["start_station_id"].value_counts()

from datetime import datetime

# Create baseline models for each location using simple mean 

ts_data_5905 = ts_data[ts_data["start_station_id"] == 5905.140137]
ts_data_6140 = ts_data[ts_data["start_station_id"] == 6140.049805]
ts_data_6822 = ts_data[ts_data["start_station_id"] == 6822.089844]

cutoff_date=datetime(2025, 1, 1, 0, 0, 0)

ts_data_5905_train = ts_data_5905[ts_data_5905["start_hour"] < cutoff_date]
ts_data_6140_train = ts_data_6140[ts_data_6140["start_hour"] < cutoff_date]
ts_data_6822_train = ts_data_6822[ts_data_6822["start_hour"] < cutoff_date]

ts_data_5905_test = ts_data_5905[ts_data_5905["start_hour"] >= cutoff_date]
ts_data_6140_test = ts_data_6140[ts_data_6140["start_hour"] >= cutoff_date]
ts_data_6822_test = ts_data_6822[ts_data_6822["start_hour"] >= cutoff_date]

ts_data_6822_test

import numpy as np

class MeanModel:
    def __init__(self):
        self.mean_value = None
        self.fitted = False

    def fit(self, ts_train: pd.DataFrame):
        """
        Fit the model by computing the mean of the time series.
        """
        if not isinstance(ts_train, pd.DataFrame):
            raise ValueError("Input must be a pandas DataFrame.")
        self.mean_value = ts_train['rides'].mean()
        self.fitted = True
        print(f"Model fitted. Mean = {self.mean_value:.2f}")

    def predict(self, ts_test: pd.DataFrame) -> np.array:
        if not self.fitted:
            raise ValueError("Model must be fitted before predicting.")
        return np.full(shape=len(ts_test), fill_value=self.mean_value)

model_5905 = MeanModel()
model_6140 = MeanModel()
model_6822 = MeanModel()

model_5905.fit(ts_data_5905_train)
model_6140.fit(ts_data_6140_train)
model_6822.fit(ts_data_6822_train)

preds_5905 = model_5905.predict(ts_data_5905_test)
preds_6140 = model_6140.predict(ts_data_6140_test)
preds_6822 = model_6822.predict(ts_data_6822_test)

from sklearn.metrics import mean_absolute_error

test_mae_5905 = mean_absolute_error(ts_data_5905_test["rides"], preds_5905)
test_mae_6140 = mean_absolute_error(ts_data_6140_test["rides"], preds_6140)
test_mae_6822 = mean_absolute_error(ts_data_6822_test["rides"], preds_6822)

print(f"MAE for 5905: {test_mae_5905:.2f}")
print(f"MAE for 6140: {test_mae_6140:.2f}")
print(f"MAE for 6822: {test_mae_6822:.2f}")

from src.experiment_utils import set_mlflow_tracking, log_model_to_mlflow
from dotenv import load_dotenv
load_dotenv()

mlflow = set_mlflow_tracking()

log_model_to_mlflow(model_5905, ts_data_5905_test, "BaselineSimpleMean5905", "mean_absolute_error", score=test_mae_5905)
log_model_to_mlflow(model_6140, ts_data_6140_test, "BaselineSimpleMean6140", "mean_absolute_error", score=test_mae_6140)
log_model_to_mlflow(model_6822, ts_data_6822_test, "BaselineSimpleMean6822", "mean_absolute_error", score=test_mae_6822)
