==> notebooks/10_lgbm_feature_importance.ipynb <==
%load_ext autoreload
%autoreload 2

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

import pandas as pd
from src.config import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")

df_5905 = df[df["start_station_id"] == 5905.140137]
df_6140 = df[df["start_station_id"] == 6140.049805]
df_6822 = df[df["start_station_id"] == 6822.089844]

from datetime import datetime

from src.data_utils import split_time_series_data

X_train_5905, y_train_5905, X_test_5905, y_test_5905 = split_time_series_data(
    df_5905,
    cutoff_date=datetime(2025, 1, 1, 0, 0, 0),
    target_column="target"
)

print(X_train_5905.shape)
print(y_train_5905.shape)
print(X_test_5905.shape)
print(y_test_5905.shape)

X_train_6140, y_train_6140, X_test_6140, y_test_6140 = split_time_series_data(
    df_6140,
    cutoff_date=datetime(2025, 1, 1, 0, 0, 0),
    target_column="target"
)

print(X_train_6140.shape)
print(y_train_6140.shape)
print(X_test_6140.shape)
print(y_test_6140.shape)

X_train_6822, y_train_6822, X_test_6822, y_test_6822 = split_time_series_data(
    df_6822,
    cutoff_date=datetime(2025, 1, 1, 0, 0, 0),
    target_column="target"
)

print(X_train_6822.shape)
print(y_train_6822.shape)
print(X_test_6822.shape)
print(y_test_6822.shape)

past_ride_columns = [c for c in X_train_5905.columns if c.startswith("rides_")]
X_train_only_numeric_5905 = X_train_5905[past_ride_columns]
X_test_only_numeric_5905 = X_test_5905[past_ride_columns]

X_train_only_numeric_6140 = X_train_6140[past_ride_columns]
X_test_only_numeric_6140 = X_test_6140[past_ride_columns]

X_train_only_numeric_6822 = X_train_6822[past_ride_columns]
X_test_only_numeric_6822 = X_test_6822[past_ride_columns]

from pathlib import Path
import pickle

path5905 = Path('..') / "models" / "lgbm_5905.pkl"
path6140 = Path('..') / "models" / "lgbm_6140.pkl"
path6822 = Path('..') / "models" / "lgbm_6822.pkl"

with open(path5905, 'rb') as f:
    model_5905 = pickle.load(f)

with open(path6140, 'rb') as f:
    model_6140 = pickle.load(f)

with open(path6822, 'rb') as f:
    model_6822 = pickle.load(f)

importances_5905 = model_5905.feature_importances_
feature_names_5905 = X_train_only_numeric_5905.columns

importance_df = pd.DataFrame({'feature': feature_names_5905, 'importance': importances_5905})
importance_df.sort_values(by='importance', ascending=False, inplace=True)

# Display top features
importance_df.head(10)

features_5905 = [
    "rides_t-1",
    "rides_t-9",
    "rides_t-11",
    "rides_t-504",
    "rides_t-2",
    "rides_t-156",
    "rides_t-13",
    "rides_t-10",
    "rides_t-414",
    "rides_t-662"
]

importances_6140 = model_6140.feature_importances_
feature_names_6140 = X_train_only_numeric_6140.columns

importance_df = pd.DataFrame({'feature': feature_names_6140, 'importance': importances_6140})
importance_df.sort_values(by='importance', ascending=False, inplace=True)

# Display top features
importance_df.head(10)

features_6140 = [
    "rides_t-11",
    "rides_t-8",
    "rides_t-653",
    "rides_t-5",
    "rides_t-2",
    "rides_t-39",
    "rides_t-28",
    "rides_t-296",
    "rides_t-55",
    "rides_t-108"
]

importances_6822 = model_6822.feature_importances_
feature_names_6822 = X_train_only_numeric_6822.columns

importance_df = pd.DataFrame({'feature': feature_names_6822, 'importance': importances_6822})
importance_df.sort_values(by='importance', ascending=False, inplace=True)

# Display top features
importance_df.head(10)

features_6822 = [
    "rides_t-1",
    "rides_t-77",
    "rides_t-5",
    "rides_t-132",
    "rides_t-18",
    "rides_t-295",
    "rides_t-78",
    "rides_t-514",
    "rides_t-269",
    "rides_t-111"
]

X_train_only_numeric_5905 = X_train_5905[features_5905]
X_test_only_numeric_5905 = X_test_5905[features_5905]

X_train_only_numeric_6140 = X_train_6140[features_6140]
X_test_only_numeric_6140 = X_test_6140[features_6140]

X_train_only_numeric_6822 = X_train_6822[features_6822]
X_test_only_numeric_6822 = X_test_6822[features_6822]

import lightgbm as lgb

model_5905 = lgb.LGBMRegressor()
model_5905.fit(X_train_only_numeric_5905, y_train_5905)

model_6140 = lgb.LGBMRegressor()
model_6140.fit(X_train_only_numeric_6140, y_train_6140)

model_6822 = lgb.LGBMRegressor()
model_6822.fit(X_train_only_numeric_6822, y_train_6822)

from sklearn.metrics import mean_absolute_error

preds_5905 = model_5905.predict(X_test_only_numeric_5905)
preds_6140 = model_6140.predict(X_test_only_numeric_6140)
preds_6822 = model_6822.predict(X_test_only_numeric_6822)

test_mae_5905 = mean_absolute_error(y_test_5905, preds_5905)
test_mae_6140 = mean_absolute_error(y_test_6140, preds_6140)
test_mae_6822 = mean_absolute_error(y_test_6822, preds_6822)

print(f"MAE for 5905: {test_mae_5905:.2f}")
print(f"MAE for 6140: {test_mae_6140:.2f}")
print(f"MAE for 6822: {test_mae_6822:.2f}")

from src.experiment_utils import set_mlflow_tracking, log_model_to_mlflow
from dotenv import load_dotenv
import os
load_dotenv()

mlflow = set_mlflow_tracking()

log_model_to_mlflow(model_5905, X_test_only_numeric_5905, "LGBMRegressor_FI_5905", "mean_absolute_error", score=test_mae_5905)
log_model_to_mlflow(model_6140, X_test_only_numeric_6140, "LGBMRegressor_FI_6140", "mean_absolute_error", score=test_mae_6140)
log_model_to_mlflow(model_6822, X_test_only_numeric_6822, "LGBMRegressor_FI_6822", "mean_absolute_error", score=test_mae_6822)
