==> notebooks/14_predict_using_hopsworks_model.ipynb <==
%load_ext autoreload
%autoreload 2

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))
import src.config as config

from src.inference import get_feature_store

from datetime import datetime, timedelta
import pandas as pd

# Get the current datetime64[us, Etc/UTC]
current_date = pd.Timestamp.now(tz="US/Eastern")
feature_store = get_feature_store()

# read time-series data from the feature store
fetch_data_to = current_date - timedelta(hours=1)
fetch_data_from = current_date - timedelta(days=1*29)
print(f"Fetching data from {fetch_data_from} to {fetch_data_to}")
feature_view = feature_store.get_feature_view(
    name=config.FEATURE_VIEW_NAME, version=config.FEATURE_VIEW_VERSION
)

ts_data = feature_view.get_batch_data(
    start_time=(fetch_data_from - timedelta(days=1)),
    end_time=(fetch_data_to + timedelta(days=1)),
)
ts_data = ts_data[ts_data.start_hour.between(fetch_data_from, fetch_data_to)]

ts_data.sort_values(["start_station_id", "start_hour"]).reset_index(drop=True)

ts_data.info()

ts_data["start_hour"] = ts_data["start_hour"].dt.tz_convert("US/Eastern")

ts_data.sort_values(["start_station_id", "start_hour"]).reset_index(drop=True)

ts_data.info()

from src.data_utils import transform_ts_data_into_features_and_target_loop

features, targets = transform_ts_data_into_features_and_target_loop(ts_data, window_size=24*28, step_size=23)

features

features.info()

from src.inference import load_batch_of_features_from_store
current_date = pd.Timestamp.now(tz='US/Eastern')
features = load_batch_of_features_from_store(current_date)

current_date

features

from src.inference import load_model_from_registry

model_5905 = load_model_from_registry(station_id=5905)
model_6140 = load_model_from_registry(station_id=6140)
model_6822 = load_model_from_registry(station_id=6822)

from src.inference import get_model_predictions

predictions_5905 = get_model_predictions(model_5905, features[features["start_station_id"] == 5905.140137])
predictions_6140 = get_model_predictions(model_6140, features[features["start_station_id"] == 6140.049805])
predictions_6822 = get_model_predictions(model_6822, features[features["start_station_id"] == 6822.089844])

predictions_5905

predictions_6140

predictions_6822

predictions = pd.concat([predictions_5905, predictions_6140, predictions_6822], ignore_index=True)
predictions.sort_values("predicted_demand", ascending=False).head(10)["start_station_id"].values
