==> notebooks/17_plot_mae.ipynb <==
%load_ext autoreload
%autoreload 2

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))
import src.config as config

from datetime import datetime, timedelta
import pytz
import pandas as pd

current_date = pd.to_datetime(datetime.now()).ceil("h").tz_localize("US/Eastern")
current_date

import hopsworks
import pandas as pd
from datetime import timedelta
from src.inference import get_feature_store, fetch_predictions

def fetch_hourly_rides(hours):
    current_hour = (pd.Timestamp.now("US/Eastern") - timedelta(hours=hours)).floor('h')

    fs = get_feature_store()
    fg = fs.get_feature_group(
        name=config.FEATURE_GROUP_NAME,
        version=1
    )

    query = fg.select_all()
    query = query.filter(fg.start_hour >= current_hour)

    return query.read()

pd.Timestamp.now(tz="US/Eastern").floor('h')

df = fetch_hourly_rides(1)

df = df.sort_values(['start_hour']).reset_index(drop=True)

df.info()

df["start_hour"] = df["start_hour"].dt.tz_convert("US/Eastern")
df

df_pred = fetch_predictions(1)

df_pred

merged_df = pd.merge(df, df_pred, on=['start_station_id', 'start_hour'])

merged_df

merged_df['difference'] = merged_df['predicted_demand'] - merged_df['rides']

merged_df.sort_values(["start_station_id", "start_hour"])

import pandas as pd
import plotly.express as px
df1 = df
df2 = df_pred

# Merge the DataFrames on 'pickup_location_id' and 'pickup_hour'
merged_df = pd.merge(df1, df2, on=['start_station_id', 'start_hour'])

# Calculate the absolute error
merged_df['absolute_error'] = abs(merged_df['predicted_demand'] - merged_df['rides'])

# Group by 'pickup_hour' and calculate the mean absolute error (MAE)
mae_by_hour = merged_df.groupby('start_hour')['absolute_error'].mean().reset_index()
mae_by_hour.rename(columns={'absolute_error': 'MAE'}, inplace=True)

# Create a Plotly plot
fig = px.line(
    mae_by_hour,
    x='start_hour',
    y='MAE',
    title='Mean Absolute Error (MAE) by Pickup Hour',
    labels={'start_hour': 'Start Hour', 'MAE': 'Mean Absolute Error'},
    markers=True
)

# Show the plot
fig.show()


mae_by_hour["MAE"].mean()
